Completely rework the way the inventory works. Currently known broken: zabbix-agent and zabbix-proxy
HACK: The 000_inventory_hook.py rewrites the inventory on the fly. This doesn't fully work for group memberships when a group membership limit is specified on the command line. That is:  ansible-playbook -i inventory/foo  works, ansible-playbook -l foo only works when the group is hardcoded in the inventory file. IOW: Doesn't work for virtual host groups (nextcloud-server, zabbix-server, etc.pp.).